Move CustomComand to plugin folder
